hector_pose_estimation_core: do not scale variance of magneto measurement in normalized mode
With this patch the magnetic/stddev parameter is considered as the standard deviation of the
normalized measurement vector and not of the raw measurement if the magnitude of the field
is not modeled (the magnetic/magnitude parameter is 0). If the magnitude is set, the normalization
is inactive and the stddev relates to the original measurement vector.