hector_pose_estimation_core: added cmake cache variable for maximum number of state variables
hector_pose_estimation only uses a maximum of 18 state variables. You should set this to a higher number to support bigger state variables,
e.g. for plugins and/or application-specific extensions.
The maximum state vector size is maximum state variables + 1 as the orientation is stored as a quaternion internally (four dimensions for three degrees of freedom).